kevm_pyk/: move kevm constructors to kevm.py
